+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Changes
 * [Developer]: Chromedriver updated to 90.
 * [Developer]: Updated `lodash` to v4.17.21 to fix security vulnerability.
 * [UI/REST]: Improved querying performance of metadata for line list and REST API.
+* [UI]: Complete upgrade of project settings to use Ant Design.
 
 20.09 to 21.01
 --------------